split out scatter.selectPoints to a new file
